requirements.txt
fastapi
uvicorn
openai==0.27.0
pydantic
python-dotenv
requests
streamlit
modal-client
